fix a logic error in parseList where it could drive off the end of the file and infinitely loop because it didn't realize that skipUntil doesn't always consume something.
This fixes an infinite loop on this testcase:

func isSpace(c : Char) -> Bool {
  return (c == '\v' ||
          c == '\f')
}

which comes from rdar://11936003.  I'm not adding it because it only works
as the last thing in a file, and isn't likely to regress.  The diagnostics
produced are also still really really awful for this.


Swift SVN r5241