Turn off the fallback to the old type checker when importing swift.swift.
With the new solver, we're only about 30% slower than the old type
checker when parsing swift.swift. That's close enough to stop using it
as an excuse to keep the old type checker around. Overall "lit" times
are about the same as before we switched to the new solver.


Swift SVN r5284