Type check the condition of ternary expressions separately.
Whether or how the condition of a ternary expression type-checks is a
separate problem from the then/else branches of the ternary
expression. Actually treat it as such.

This eliminates the last place in the frontend where we mode-switch
into the old type checker intentionally. It's still used in various
places externally in the test suite (via -no-constraint-checker) and
implicit (e.g., via calls to semaApplyExpr).


Swift SVN r5285