SILGen: Map String to NSString in ObjC method thunks.
If a method or property on an [objc] class takes String arguments or returns a String result in the Swift world, insert conversions to/from NSString into the ObjC thunk for the method.

Moving the code around in emitObjCThunk here induces a bit of extra retain/release traffic. But the SIL optimizer will take care of that real soon, right?

Swift SVN r5359