SILGen: Introduce writeback scopes.
Define a new WritebackScope RAII object to mark the beginning and end of the lifetime of a set of writebacks. Add a writeback stack to SILGenFunction to keep track of active writebacks. Replace the ad-hoc writeback tracking in emitAssignToLValue with a WritebackScope.

No functionality change just yet; this just sets the stage for handling writebacks in function arguments.

Swift SVN r5402