SILGen: Introduce writeback scopes for calls.
When emitting a call chain, wrap the natural uncurry level of the root call and subsequent chained calls in their own writeback scopes. Suppress writeback if a non-settable property is used as the 'this' argument of a method and during codegen of LoadExprs.

This uncovered a bug in SILGen where a subscript LValue reevaluates the subscript expression on every access, which breaks the stdlib implementation of printf(). I've applied a temporary workaround to keep it working; I'll fix it next.

Swift SVN r5410