Hack-and-slash porting parts of Clang's serialized diagnostic printer over to Swift.
This is still total WIP.  There's a bunch of copy-paste, both from other
parts of Swift and Clang.  I want to see the end-to-end vision
first and then refactor.  There's a bunch of subtle changes between
Clang's diagnostics and Swift's that require some care to handle,
as the serialized diagnostics were developed with Clang's in mind.
That should be fine, since Clang' will be more complicated, but the
differences are still there to be dealt with.

It's not clear if we will need the SharedState, as we do with Clang.
We'll keep it now since it isn't doing any harm.

Swift SVN r5507