AST: Save ProtocolConformances in TypeDecls and ExtensionDecls.
When type-checking explicit protocol conformances for a type or extension, save the ProtocolConformance mappings from that declaration in the decl node. We'll need this in order to emit ObjC protocol conformances, which can't be emitted internally on-demand the way we currently emit Swift protocol witness tables. No functionality change.

Swift SVN r5632