[Constraint solver] Don't guess at type variable bindings that are guaranteed to be fully bound by other constraints.
A type variable can be fully bound by either an overload set (when
selecting an overload will provide a binding for the type variable) or
by a member constraint (when the base type is determined, the member
type will be known). This keeps the solver from making really, really
bad type variable bindings based on context information.


Swift SVN r5634