Sema: Propagate ObjC-ness of protocol mappings at type-checking time.
Tag methods as 'isObjC' during decl type-checking if they're mapped into an ObjC protocol conformance. This is better than r5646 because it means we don't need another pass of thunking in SILGen and IRGen, and the thunks will get mapped in a deterministic order and not freak out the buildbot.

Swift SVN r5654