REPL: Fix printing of containers embedded in structs.
Apply the current MemberIndexes path for a collection at the top level, and not try to reuse that path relative to the (key, value) bindings of the loop body. When we build the expression for a MemberIndexes path, type-check each step as we go so we don't crash on a null Type trying to discriminate each path element of a struct in a struct in a struct. Fixes <rdar://problem/14211671>.

Swift SVN r5725