Revise the weak design in response to comments.
I actually backtracked on the name and went back to [backref].

The main changes here are (1) [backref] is now definitively a
decl attribute and (2) there's now a huge new section about
closures.  This entailed doing quite a bit of secondary design
on closure capture semantics because backref captures are
value captures, and adding those just does so much fun stuff
to the model.

Swift SVN r5836