[stdlib] Replace String storage with a reference-counted buffer with capacity
Allows us to use a copy-on-write strategy to efficiently grow Strings.
This should obviate the need for a StringBuffer class anytime soon and
make comfortable formatting efficient.

If we backslide and the COW/capacity functionality of String is no
longer doing its job, the test added here will appear to hang and
hopefully someone will be annoyed enough to fix it.  We can't do
better right now because performance measurements vary based on the
test platform.

Swift SVN r5845