SILGen: Emit case block BBs as part of dispatch.
Put emitDecisionTree in charge of setting up case-to-BB mappings as they are needed by leaves of the decision tree. This lets us avoid emitting empty waypoint BBs for default dispatches, cull unreachable case blocks, and will let us handle the scoping of pattern variables relative to case blocks. This new implementation breaks fallthrough, but it's going to need reworking in the face of case variables anyway, so tear out the old implementation and mark it as not implemented for now.

Swift SVN r6094